✨ Feat :  사진 저장을 위한 권한 모달 텍스트 공통 유틸로 승격
