+++ src/shared/constants/text/photoPermissionText.ts
@@ -1,0 +1,6 @@
+export const SAVE_PERMISSION_MODAL = {
+  TITLE: '설정에서 픽셀의\n사진 접근을 허용해야 해요',
+  DESCRIPTION: '설정 > 픽셀 > 사진에서\n사진 접근 허용으로 변경해 주세요',
+  CANCEL: '취소',
+  CONFIRM: '설정으로 이동',
+} as const;
